server/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
scikit-learn
umap-learn
numpy<=2.1
python-dotenv